+++ interface/src/App/PvP/PvpPage.jsx
@@ -24,7 +24,8 @@ import { MovePoolBuilder } from "js/movePoolBuilder";
 import { separateMovebase } from "js/separateMovebase";
 import { returnPokList } from "js/returnPokList";
 
-import { calculateMaximizedStats, calculateEffStat } from "js/indexFunctions";
+import { calculateEffStat } from "js/indexFunctions";
+import { calculateMaximizedStats } from "js/Maximizer/Maximizer";
 import { extractData } from "js/extractors/extractData";
 import { extractPokemon } from "js/extractors/extractPokemon";
 import { getCookie } from "js/getCookie";
